+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ## Changed
 - artists now gets turbo colormap straight from matplotlib
 - deprecating  `artists.plot_colorbar`: instead use matplotlib's `colorbar` implementations directly
+- artists.interact2D now returns a `types.SimpleNamespace` object (rather than a tuple)
 
 ## [3.5.2]